middleware/querylog/middleware_test.go
package querylog

import (
	"context"
	"database/sql"
	"testing"

	"gitee.com/youkelike/orm"
	_ "github.com/mattn/go-sqlite3"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
)

func TestMiddlewareBuilder(t *testing.T) {
	var query string
	var args []any
	m := (&MiddlewareBuilder{}).LogFunc(func(q string, as []any) {
		query = q
		args = as
	})

	db, err := orm.Open("sqlite3", "file:test.db?cache=shared&mode=memory", orm.DBWithMiddlewares(m.Build()))
	require.NoError(t, err)
	_, _ = orm.NewSelector[TestModel](db).Where(orm.C("Id").Eq(10)).Get(context.Background())
	assert.Equal(t, "SELECT * FROM test_model WHERE id=?;", query)
	assert.Equal(t, []any{10}, args)

	_ = orm.NewInserter[TestModel](db).Values(&TestModel{Id: int64(18)}).Exec(context.Background())
	assert.Equal(t, "INSERT INTO test_model (id,first_name,age,last_name) VALUES (?,?,?,?);", query)
	assert.Equal(t, []any{int64(18), "", int8(0), (*sql.NullString)(nil)}, args)
}

type TestModel struct {
	Id        int64
	FirstName string
	Age       int8
	LastName  *sql.NullString
}
